+++ packages/apollo-link-batch-http/src/batchHttpLink.ts
@@ -129,8 +129,12 @@ export class BatchHttpLink extends ApolloLink {
       }
 
       return new Observable<FetchResult[]>(observer => {
-        // the raw response is attached to the context in the BatchingLink
         fetcher(chosenURI, options)
+          .then(response => {
+            // Make the raw response available in the context.
+            operations.forEach(operation => operation.setContext({ response }));
+            return response;
+          })
           .then(parseAndCheckHttpResponse(operations))
           .then(result => {
             // we have data and can send it to back up the link chain