+++ packages/apollo-link-http/src/__tests__/sharedHttpTests.ts
@@ -567,6 +567,35 @@ export const sharedHttpTest = (
         }),
       );
     });
+
+    it('sets the raw response on context', done => {
+      const middleware = new ApolloLink((operation, forward) => {
+        return new Observable(ob => {
+          const op = forward(operation);
+          const sub = op.subscribe({
+            next: ob.next.bind(ob),
+            error: ob.error.bind(ob),
+            complete: makeCallback(done, e => {
+              expect(operation.getContext().response.headers.toBeDefined);
+              ob.complete();
+            }),
+          });
+
+          return () => {
+            sub.unsubscribe();
+          };
+        });
+      });
+
+      const link = middleware.concat(createLink({ uri: 'data', fetch }));
+
+      execute(link, { query: sampleQuery }).subscribe(
+        result => {
+          done();
+        },
+        () => {},
+      );
+    });
   });
 
   describe('dev warnings', () => {